Spell out the swipe-fade types so every Swift version accepts them
CI rejected `1 - min(abs(dismissOffset) / 180, 0.45)` as an ambiguous
use of `/` while the same expression built locally — the literals leave
the type checker a choice, and which way it goes depends on the Swift
version. Naming CGFloat explicitly removes the choice.

+++ NotchPill/Views/Tiles.swift
@@ -590,7 +590,10 @@ struct DevReadyPeekRow: View {
             // Waiting rows intentionally ignore this gesture: an approval must
             // always require the explicit × or Esc dismissal path.
             .offset(x: dismissOffset)
-            .opacity(1 - min(abs(dismissOffset) / 180, 0.45))
+            // Spelled out in CGFloat rather than left to inference: the
+            // literals here are ambiguous enough that some Swift versions
+            // reject the expression outright.
+            .opacity(Double(1 - min(abs(dismissOffset) / CGFloat(180), CGFloat(0.45))))
             .simultaneousGesture(dismissGesture)
             .accessibilityHint(alert.kind == .finished
                                ? "Swipe left to dismiss, or double tap to open"
